+++ src/App.test.js
@@ -1,8 +1,8 @@
 import { render, screen } from '@testing-library/react';
 import App from './App';
 
-test('renders learn react link', () => {
+test('Show the introductory IIOT Test Bench header', () => {
   render(<App />);
-  const linkElement = screen.getByText(/learn react/i);
+  const linkElement = screen.getByText(/IIOT Test Bench/i);
   expect(linkElement).toBeInTheDocument();
 });
